+++ docs/CHANGELOG.rst
@@ -14,6 +14,7 @@ Unreleased
 Added
 -----
 - Add ALT aware alignment option to BWA-MEM process
+- Extend the documentation with examples of writing workflows and tests
 
 Changed
 -------